# Hustle-Safe/Hustle-Safe/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HustleSafe API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /zones/{id}:
    get:
      operationId: getZone
      tags: [zones]
      summary: Get a zone by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Zone details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"
    patch:
      operationId: updateZone
      tags: [zones]
      summary: Update zone data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateZoneRequest"
      responses:
        "200":
          description: Updated zone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"
...
components:
  schemas:
...
    Zone:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        city:
          type: string
        lat:
          type: number
          nullable: true
        lng:
          type: number
          nullable: true
        gds_score:
          type: integer
        status:
          type: string
          enum: [normal, elevated, high, shutdown]
        active_workers:
          type: integer
        rainfall_mm:
          type: number
        traffic_score:
          type: number
        aqi:
          type: integer
        demand_drop_pct:
          type: integer
        govt_alert:
          type: boolean
        last_updated:
          type: string
      required:
        - id
        - name
        - city
        - gds_score
        - status
        - active_workers
        - rainfall_mm
        - traffic_score
        - aqi
        - demand_drop_pct
        - govt_alert
        - last_updated
...
    UpdateZoneRequest:
      type: object
      properties:
        gds_score:
          type: integer
        status:
          type: string
          enum: [normal, elevated, high, shutdown]
        rainfall_mm:
          type: number
        traffic_score:
          type: number
        aqi:
          type: integer
        demand_drop_pct:
          type: integer
        govt_alert:
          type: boolean
        active_workers:
          type: integer
...